+++ generators/dependencies.js
@@ -2,13 +2,14 @@
 // TO-DO: maybe add as peerDependencies for automated security checks via GitHub, then read programmatically from package.json
 export default {
 	"@sap-ux/eslint-plugin-fiori-tools": "^0.2",
-    "@sap/approuter": "^16",
-    "bestzip": "latest",
-    "mbt": "^1",
-    "npm-run-all": "^4",
-    "rimraf": "latest",
-    "OpenUI5": "1.120.13",
-    "SAPUI5": "1.120.13",
+	"@sap/approuter": "^16",
+	"@ui5/linter": "latest",
+	"bestzip": "latest",
+	"mbt": "^1",
+	"npm-run-all": "^4",
+	"rimraf": "latest",
+	"OpenUI5": "1.120.13",
+	"SAPUI5": "1.120.13",
 	"ui5-middleware-approuter": "^3",
 	"ui5-task-nwabap-deployer": "^2"
 }
