+++ generators/project/templates/package.json
@@ -5,7 +5,8 @@
         "build": "mbt build --mtar <%= title %>.mtar",
         "build:workspaces": "npm run build --workspaces --if-present",
         "deploy": "cf deploy mta_archives/<%= title %>.mtar --retries 0",
-		"lint": "npm run lint --workspaces --if-present"
+		"lint": "npm run lint --workspaces --if-present",
+		"ui5lint": "npm run ui5lint --workspaces --if-present"
     },
     "dependencies": {
         "mbt": "<%= mbtVersion %>"
